Moved region inference error reporting into own module.
